Apply debate-consensus refactors to Cloud.Platforms
Design-review follow-ups (multi-agent consensus):
- LinkExtensions.ToUri: guard the parse (Uri.TryCreate) so relative internal
  links return null instead of throwing UriFormatException.
- GetContentUrls: extract to GetContentUrlsQuery + GetContentUrlsHandler
  (mirrors GetPreviewHtml); controller is now a thin mediator delegate with a
  single IMediator dependency.
- ContentUrlsRes: drop the redundant Permitted flag (both-null already means
  "no URLs"); client reads stagingUrl/productionUrl directly.
- GetPreviewHtmlHandler: drop redundant IContentTypeService; use the loaded
  IContent's ContentType.Alias directly. Delete orphaned NamedParameter
  ContentTypeAlias.
- Preview visibility: replace custom PlatformsPreviewVisibilityController +
  N3O.Condition.WorkspaceVisibility with the first-party
  Umb.Condition.WorkspaceContentTypeAlias (match: platformsOffering); delete
  the controller.
- crowdfunding-visibility: tighten note detection from includes() to
  endsWith() to avoid alias false-positives.
- DonationFormStateValidator.Fund: remove dead second .OrEmpty().
- Remove dead comment-only CampaignSending.cs / OfferingSending.cs stubs
  (migration notes belong in the migration docs on the integration branch).
