--- src/Cloud/N3O.Umbraco.Cloud.Platforms/NamedParameters/ContentTypeAlias.cs
@@ -1,7 +1,0 @@
-﻿using N3O.Umbraco.Parameters;
-
-namespace N3O.Umbraco.Cloud.Platforms.NamedParameters;
-
-public class ContentTypeAlias : NamedParameter<string> {
-    public override string Name => "contentTypeAlias";
-}
