--- src/Cloud/N3O.Umbraco.Cloud.Platforms/Notifications/Offerings/OfferingSending.cs
@@ -1,5 +1,0 @@
-// v17 status (Bellissima migration):
-//
-// Staging/production URL display for offerings: IMPLEMENTED as the N3O.WorkspaceView.PlatformsUrls
-// workspace view tab (platforms-urls.ts / PlatformsBackOfficeController.GetContentUrls).
-namespace N3O.Umbraco.Cloud.Platforms.Notifications;
