+++ src/Cloud/N3O.Umbraco.Cloud.Platforms/Validators/DonationFormState/DonationFormStateValidator.Fund.cs
@@ -47,7 +47,6 @@ public class FundDonationFormStateValidator : DonationFormStateValidator<FundDon
         var suggestedAmounts = property.IfNotNull(x => ContentHelper.GetBlockList(x))
                                        .OrEmpty()
                                        .Select(x => x.Content)
-                                       .OrEmpty()
                                        .As<DonationFormStateSuggestedAmountElement>()
                                        .ToList();
